Added BuildingInfo tags. * Adjusted RenX_ProcessTags in RenX::Plugin accordingly. ** Updated Plugins accordingly. Changed BuildingInfo health/max_health type to double Added Rx_Building_ translations. Fixed binfo RCON command spam, and improved Server::updateClientList(). Updated Jupiter.

+++ RenX.Core/RenX_BuildingInfo.h
@@ -43,8 +43,8 @@ namespace RenX
 	{
 		Jupiter::StringS name;
 		TeamType team = TeamType::Other;
-		int health = 0;
-		int max_health = 0;
+		double health = 0;
+		double max_health = 0;
 		bool capturable = false;
 		mutable Jupiter::INIFile varData;
 	};
